+++ src/Command/DownloadWeburgCommand.php
@@ -3,11 +3,11 @@
 namespace Popstas\Transmission\Console\Command;
 
 use GuzzleHttp;
+use Popstas\Transmission\Console\WeburgClient;
 use Symfony\Component\Console\Helper\ProgressBar;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Console\Logger\ConsoleLogger;
 
 class DownloadWeburgCommand extends Command
 {
@@ -27,12 +27,15 @@ EOT
     protected function execute(InputInterface $input, OutputInterface $output)
     {
         $logger = $this->getLogger($output);
+        $httpClient = new GuzzleHttp\Client();
+        $weburgClient = new WeburgClient($httpClient);
+
         $torrents_dir = $input->getOption('dest');
         if (!$torrents_dir) {
             $torrents_dir = $this->config->get('download-torrents-dir');
         }
         if (!$torrents_dir) {
-            $output->writeln('<error>Download destination direcrory not set.</error>');
+            $output->writeln('<error>Download destination directory not set.</error>');
             $output->writeln('Use command with --dest=/path/to/dir parameter '
                 .'or define destination directory in config file.');
             exit(1);
@@ -43,7 +46,7 @@ EOT
             mkdir($download_dir, 0777, true);
         }
 
-        $movies_ids = $this->getMoviesIds();
+        $movies_ids = $weburgClient->getMoviesIds();
 
         $progress = new ProgressBar($output, count($movies_ids));
         $progress->start();
@@ -59,12 +62,20 @@ EOT
                 continue;
             }
 
-            $movie_info = $this->getTorrentInfo($movie_id);
+            $movie_info = $weburgClient->getMovieInfoById($movie_id);
 
-            if ($this->isTorrentPopular($movie_info)) {
+            $isTorrentPopular = $this->isTorrentPopular(
+                $movie_info,
+                $this->config->get('download-comments-min'),
+                $this->config->get('download-imdb-min'),
+                $this->config->get('download-kinopoisk-min'),
+                $this->config->get('download-votes-min')
+            );
+
+            if ($isTorrentPopular) {
                 $progress->setMessage('Download movie ' . $movie_id . '...');
 
-                $torrents_urls = $this->getTorrentUrls($movie_id, $logger);
+                $torrents_urls = $weburgClient->getMovieTorrentUrlsById($movie_id);
                 if (!$input->getOption('dry-run')) {
                     $this->downloadTorrents($torrents_urls, $torrents_dir);
 
@@ -81,93 +92,12 @@ EOT
         $progress->finish();
     }
 
-    private function getUrlBody($url, $headers = [])
+    private function isTorrentPopular($movie_info, $comments_min, $imdb_min, $kinopoisk_min, $votes_min)
     {
-        $client = new GuzzleHttp\Client();
-        $jar = new GuzzleHttp\Cookie\CookieJar();
-
-        $res = $client->request('GET', $url, [
-            'headers' => $headers,
-            'cookies' => $jar,
-        ]);
-
-        if (!$res->getStatusCode()) {
-            echo 'error ' . $res->getStatusCode();
-            exit(1);
-        }
-
-        return $res->getBody();
-    }
-
-    private function getMoviesIds()
-    {
-        $movie_id_regex = '\/movies\/info\/([0-9]+)';
-        $movies_url = 'http://weburg.net/movies/new/?clever_title=1&template=0&last=0';
-
-        $json_raw = $this->getUrlBody($movies_url, [
-            'Content-Type'     => 'text/html; charset=utf-8',
-            'User-Agent'       => 'Mozilla/5.0 (X11; Ubuntu; Linux x86_64; rv:27.0) Gecko/20100101 Firefox/27.0',
-            'X-Requested-With' => 'XMLHttpRequest',
-        ]);
-
-        $movies_json = json_decode($json_raw);
-        preg_match_all('/' . $movie_id_regex . '/', $movies_json->items, $res);
-        $movies_ids = array_unique($res[1]);
-
-        return $movies_ids;
-    }
-
-    private function getTorrentInfo($movie_id)
-    {
-        $movie_url = 'http://weburg.net/movies/info/' . $movie_id;
-
-        $info = array();
-
-        $body = $this->getUrlBody($movie_url);
-        $body = iconv('WINDOWS-1251', 'UTF-8', $body);
-
-        preg_match('/Комментариев:&nbsp;(\d+)/', $body, $res);
-        $info['comments'] = count($res) ? $res[1] : 0;
-
-        preg_match('/external-ratings-link_type_kinopoisk.*?>(\d+\.\d+)/mis', $body, $res);
-        $info['rating_imdb'] = count($res) ? $res[1] : 0;
-
-        preg_match('/external-ratings-link_type_imdb.*?>(\d+\.\d+)/mis', $body, $res);
-        $info['rating_kinopoisk'] = count($res) ? $res[1] : 0;
-
-        preg_match('/count-votes" value="([0-9]+)"/mis', $body, $res);
-        $info['rating_votes'] = count($res) ? $res[1] : 0;
-
-        if (!$info['comments'] || !$info['comments'] || !$info['comments'] || !$info['comments']) {
-            printf("Cannot find all information about movie %s\n%s", $movie_url);
-            print_r($info);
-        }
-
-        return $info;
-    }
-
-    private function getTorrentUrls($movie_id, ConsoleLogger $logger)
-    {
-        $torrent_url = sprintf('http://weburg.net/ajax/download/movie?obj_id=%d', $movie_id);
-
-        $body = $this->getUrlBody($torrent_url);
-
-        preg_match_all('/(http:\/\/.*?)"/', $body, $res);
-        $torrents_urls = $res[1];
-
-        foreach ($torrents_urls as $torrents_url) {
-            $logger->debug('found torrent url: ' . $torrents_url);
-        }
-
-        return $torrents_urls;
-    }
-
-    private function isTorrentPopular($movie_info)
-    {
-        return $movie_info['comments'] >= $this->config->get('download-comments-min')
-        || $movie_info['rating_imdb'] >= $this->config->get('download-imdb-min')
-        || $movie_info['rating_kinopoisk'] >= $this->config->get('download-kinopoisk-min')
-        || $movie_info['rating_votes'] >= $this->config->get('download-votes-min');
+        return $movie_info['comments'] >= $comments_min
+        || $movie_info['rating_imdb'] >= $imdb_min
+        || $movie_info['rating_kinopoisk'] >= $kinopoisk_min
+        || $movie_info['rating_votes'] >= $votes_min;
     }
 
     private function downloadTorrents($torrents_urls, $torrents_dir)